refactor: change directory name, add path to variable in game __settings__.py
