--- __settings__.py
@@ -1,14 +1,0 @@
-import string
-upper_letters = string.ascii_uppercase
-letters = string.ascii_letters
-
-hovered = {
-    "start_button" : False,
-    "score_button" : False,
-    "words_button" : False,
-    "quit_button" : False,
-    "reset_button" : False,
-    "back_button" : False,
-    "left" : False,
-    "right" : False
-}
