+++ assets/graphics/game_fg_filter.png
