+++ assets/graphics/game_fg_lost.png
