+++ assets/graphics/game_mg_gallowstand.png
