+++ assets/graphics/game_mg_gallowstand_won.png
