+++ assets/graphics/game_mg_graveyard.png
