+++ assets/graphics/game_mg_hang0.png
