+++ assets/graphics/game_mg_hang1.png
