+++ assets/graphics/game_mg_hang2.png
