+++ assets/graphics/game_mg_hang3.png
