+++ assets/graphics/game_mg_hang4.png
