+++ assets/graphics/game_mg_hang5.png
