+++ assets/graphics/game_mg_hang6.png
