+++ assets/graphics/game_mg_hang7.png
