+++ assets/graphics/game_mg_mountains.png
