+++ assets/graphics/main_fg_dialog.png
