+++ assets/graphics/main_fg_left_arrow_white.png
