+++ assets/graphics/main_fg_right_arrow_red.png
