+++ assets/graphics/main_fg_right_arrow_white.png
