+++ assets/graphics/main_mg_village.png
