+++ display/display.py
@@ -1,7 +1,8 @@
-import pygame_module.__settings__ as settings
-import game_module.scores.manage_scores as scores_display
-from pygame_module.__settings__ import color, graphics, sfx
+import display.__settings__ as settings
+import game.scores.manage_scores as scores_display
+from display.__settings__ import color, graphics, sfx
 import pygame
+
 def pygame_init():
     pygame.init()
     screen = pygame.display.set_mode((640,480))
@@ -9,9 +10,11 @@ def pygame_init():
     pygame.display.set_icon(graphics['icon'])
     clock = pygame.time.Clock()
     return screen, clock
+
 def pygame_mixer(music_name):
     pygame.mixer.music.load(sfx[music_name])
     pygame.mixer.music.play(-1,0,3)
+
 def menu_button(screen, button_id, button_x, button_y, font_size=48, hovered = False):
     button_rect = settings.pixeled_dialog_text(font_size,button_id)\
         .get_rect(bottomleft=(button_x,button_y))
@@ -19,6 +22,7 @@ def menu_button(screen, button_id, button_x, button_y, font_size=48, hovered = F
         screen.blit(settings.pixeled_dialog_text(font_size,button_id,color['dark_red']), button_rect)
     else: screen.blit(settings.pixeled_dialog_text(font_size,button_id,color['white']), button_rect)
     return button_rect
+
 def main_menu(screen):
     screen.blit(graphics['main_background_clouds'],(0,0))
     screen.blit(graphics['main_middleground_village'],(0,0))
@@ -28,6 +32,7 @@ def main_menu(screen):
                 (112,'Le Pendu',color['dark_gray']), (main_title_rect[0]+2,main_title_rect[1]+2))
     screen.blit(settings.pixeled_dialog_text\
                 (112,'Le Pendu',color['dark_red']), main_title_rect)
+    
 def main_menu_buttons(screen, start_button_hovered, score_button_hovered, words_button_hovered, quit_button_hovered):
     start_button_rect = menu_button(screen,'Jouer',350,250,48,start_button_hovered)
     score_button_rect = menu_button(screen,'Score',350,315,48,score_button_hovered)
@@ -38,17 +43,20 @@ def main_menu_buttons(screen, start_button_hovered, score_button_hovered, words_
                     "words" : words_button_rect,
                     "quit" : quit_button_rect}
     return menu_buttons
+
 def get_scores_menu_pages():
     scores = scores_display.load_scores()
     players_id = list(scores.keys())
     pages = (len(players_id)//6)+(1 if len(players_id)%6!=0 else 0)
     return pages
+
 def scores_menu_blank(screen):
     screen.blit(graphics['main_background_clouds'],(0,0))
     screen.blit(graphics['main_middleground_village'],(0,0))
     blank_dialog_rect = settings.pixeled_dialog_text(32,'Aucun score enregistre')\
         .get_rect(center = (320,240))
     screen.blit(settings.pixeled_dialog_text(32,'Aucun score enregistre'),blank_dialog_rect)
+
 def scores_menu(screen,page,pages,left_hovered, right_hovered):
     screen.blit(graphics['main_background_clouds'],(0,0))
     screen.blit(graphics['main_middleground_village'],(0,0))
@@ -60,6 +68,7 @@ def scores_menu(screen,page,pages,left_hovered, right_hovered):
     player_scores(screen,players_id,scores, page)
     left_arrow_rect, right_arrow_rect = scores_menu_arrows(screen, pages, page, left_hovered, right_hovered)
     return left_arrow_rect, right_arrow_rect
+
 def player_scores(screen,players_id,scores, page):
     x =(120,320,520)
     y =(120,280)
@@ -86,6 +95,7 @@ def player_scores(screen,players_id,scores, page):
                 screen.blit(player_info,player_info_rect)
     except IndexError:
         return
+    
 def scores_menu_arrows(screen, pages, page, left_hovered, right_hovered):
     if pages > 1 and page >= 1:
         screen.blit(settings.main_arrows(0, left_hovered),(0,0))
@@ -96,6 +106,7 @@ def scores_menu_arrows(screen, pages, page, left_hovered, right_hovered):
         right_arrow_rect = pygame.Rect(340,418,44,32)
     else : right_arrow_rect = pygame.Rect(-340,418,44,32)
     return left_arrow_rect, right_arrow_rect
+
 def game_set_up_menu(screen, user_input, game_mode, dialog, is_add_word=None, new_word=None):
     screen.blit(graphics['main_dialog_box'],(0,-20))
     dialog_input_rect = settings.pixeled_dialog_text(32,dialog)\
@@ -125,6 +136,7 @@ def game_set_up_menu(screen, user_input, game_mode, dialog, is_add_word=None, ne
     screen.blit(settings.pixeled_dialog_text(32,dialog), dialog_input_rect)
     screen.blit(settings.pixeled_dialog_text(32,user_input),user_input_rect)
     return easy_button_rect, hard_button_rect
+
 def game_environment(screen,life_count, user_word_format):
     screen.blit(settings.game_background(life_count),(0,0))
     if life_count > 3:
@@ -135,6 +147,7 @@ def game_environment(screen,life_count, user_word_format):
     screen.blit(settings.game_hangedman(life_count),(0,0))
     if life_count < 3:
         screen.blit(graphics['game_foreground_filter'],(0,0))
+
 def game_interface(screen, guess_word, player_input, letters_played):
     screen.blit(graphics['main_dialog_box'],(0,60))
     guess_word_rect = settings.pixelplay_dialog_text(30)\
@@ -149,6 +162,7 @@ def game_interface(screen, guess_word, player_input, letters_played):
     screen.blit(settings.pixeled_dialog_text(30,'votre lettre :'),dialog_input_rect)
     screen.blit(settings.pixelplay_dialog_text(30,' '.join(letters_played),color['dark_red']),letters_played_rect)
     screen.blit(settings.pixelplay_dialog_text(42,player_input), player_input_rect)
+
 def end_message(screen, win_or_lose, guess_word):
     end_message_list = (
         graphics['game_dialog_lost'],
