+++ game/words/words_easy.txt
@@ -93,3 +93,4 @@ banane
 fiji
 lapinou
 lapin
+pain
