--- pygame_module/__init__.py
@@ -1,3 +1,0 @@
-import string
-upper_letters = string.ascii_uppercase
-letters = string.ascii_letters
